Move phone routing out of the ntfy transport.
